// src/app/admin/model/enumeration/CategoryType.ts
export enum CategoryType {
  EDUCATIONAL_LEVEL = 'educationalLevels',
  EDUCATIONAL_SUBJECT = 'educationalSubjects',
  ORGANIZATION = 'organizations',
}

export enum ActivityType {
  EDIT = 0,
  DOWNLOAD = 1,
  SEARCH = 2,
  VIEW = 3,
}

export enum IntervalType {
  DAY = 0,
  WEEK = 1,
  MONTH = 2,
}